Refactored the metadata classes
