+++ lib/geo_pattern/background.rb
@@ -1,10 +1,12 @@
 module GeoPattern
   class Background
+    include Roles::ComparableMetadata
+
     extend Forwardable
 
     attr_reader :image, :preset, :color
 
-    def_delegators :@preset, *Helpers.build_arguments(:base_color, :color)
+    def_delegators :@preset, :base_color, :color
 
     def initialize(options)
       @image     = options[:image]
@@ -12,5 +14,7 @@ module GeoPattern
       @color     = options[:color]
       @generator = options[:generator]
     end
+
+    def_comparators :base_color, :color
   end
 end
